requirement.txt
amqp==5.0.2
anyjson==0.3.3
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.2
django-allauth==0.42.0
django-bootstrap4==2.3.1
django-filter==2.4.0
django-seed==0.2.2
django-summernote==0.8.11.6
django-taggit==1.3.0
django-tof==0.3.2
djangorestframework==3.12.2
Faker==4.14.2
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.0.1
prompt-toolkit==3.0.8
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0